arm/arm64: perf: Fix incorrect return type for perf_misc_flags
The return value of perf_misc_flags is assigned to the `misc`
field of the perf_event_header structure, and its type should
be `unsigned short`.

Fixes: e50c541 ("ARM: perf: add guest vs host discrimination")
Fixes: 75e4246 ("arm64: perf: add guest vs host discrimination")

+++ arch/arm/include/asm/perf_event.h
@@ -11,7 +11,7 @@
 #ifdef CONFIG_PERF_EVENTS
 struct pt_regs;
 extern unsigned long perf_instruction_pointer(struct pt_regs *regs);
-extern unsigned long perf_misc_flags(struct pt_regs *regs);
+extern unsigned short perf_misc_flags(struct pt_regs *regs);
 #define perf_misc_flags(regs)	perf_misc_flags(regs)
 #endif
