+++ arch/arm64/kernel/perf_callchain.c
@@ -47,10 +47,10 @@ unsigned long perf_instruction_pointer(struct pt_regs *regs)
 	return instruction_pointer(regs);
 }
 
-unsigned long perf_misc_flags(struct pt_regs *regs)
+unsigned short perf_misc_flags(struct pt_regs *regs)
 {
 	unsigned int guest_state = perf_guest_state();
-	int misc = 0;
+	unsigned short misc = 0;
 
 	if (guest_state) {
 		if (guest_state & PERF_GUEST_USER)
